Support descendent chunks
This supports descendent chunks of manifest entrypoint chunks. The major changes are detailed below.

## Adds “chunk” to package terminology

This is consistent with Vite’s terminology in [their manifest parser](https://github.com/vitejs/vite/blob/61ea32056048e902ca69d88e1b0a2d21660dae2a/packages/vite/src/node/plugins/manifest.ts). A manifest consists of chunks (some of which my be entrypoints).

As part of this transition, serveral refactors were made across the codebase. This includes deprecating the existing `%Entry{}` in favor of `%Chunk{}`

## Adds concept of “dependent chunks”

In the manifest, it’s possible for an entrypoint to import a module that also has its own css. We now load those CSS since Vite does not dynamically load them in the import.

## Transitions top-level `Vite` to be a coordinator

I found the `Vite.View` module to be difficult to test with the above changes. I believe its focus should be mostly on presentational concerns with the core data type (`%Vite.Chunk{}`). I simplified its API to just accept chunks and returns the appropriate HTML. The top-level `Vite` module is now responsible for coordinating more of the logic.

## Pass around opts

Instead of modifying a global config, we transition to functions that accept options and fallback to the default provided by config. This makes it easier to test and override if necessary.

## Minimize `Vite.Manifest` API surface area

I’m not entirely sure what some of these functions are for. It seems like they’re for convenience? I removed them because they were added surface area for me to fix to get working.